// arunikacoffe/types.ts
export interface Product {
  id: number;
  name: string;
  description: string;
  price: string;
  image: string;
}

export interface BlogItem {
  id: number;
  title: string;
  excerpt: string;
  date: string;
  image: string;
}

export interface Category {
  id: number;
  title: string;
  image: string;
}